general-classification/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

torch==2.8.0+cpu
torchvision==0.23.0+cpu
fastapi==0.116.1
uvicorn[standard]==0.35.0
numpy==1.26.4
pillow==11.3.0
ultralytics==8.3.170
pika==1.3.2
requests==2.32.5
